Created a Routine object and a test for the Routine
Routine is for now a way to write function that yields as in some other
languages, it is not necessary and might not be but I had this idea
and could not let it go. So here they are with their fugly macros and stuff
They might be off use to create an actor model or for some high logic code.